Check in package-lock.
This is a new feature of npm that specifies exact package versions. It
gets generated every time you run `npm install` (or `make` or
`make basic`), so we need to either check this in or ignore it.
Checking it in is helpful because it avoid surprise version differences
between users.