stop trying to process keyboard actions on blur
We are seeing cases where blur is tricking mathquill into thinking there was a keypress. The hypothesis is that mathquill missed a character in the textarea for some reason (or failed to clear it) and didn't check again until blur. At that point it finds a character in the textarea and thinks it was a typed character. This then causes a dispatch-in-dispatch error because a completely unrelated dispatch was the cause of the blur.

I can't think of a time when we'd rather read a character out late compared to just ignoring the character. I think when it comes to keyboard input it's always safer to miss characters than make characters up. We should let this soak for a while to make sure this doesn't cause issues. I'm hoping to see a noticeable decrease in dispatch-in-dispatch errors over the next few days.