Separate browser and CommonJS builds
Because the immediate-mode browser use case and the CommonJS `require` use case have different needs in terms of templating, the simplest solution (for now) is to simply create multiple build artifacts that are substantially the same internally.

This change represents the differences between these use cases with different `src/entry/*.js` fiels, which contain the "wrapper" templating that would apply in each case.  These each have a `{SOURCE}` placeholder, which is replaced with the concatenated source content at build time.