Remove a load-time dependency on `window`
Since CommonJS won't necessarily be loaded in a browser environment, it's worthwhile to cleanup code that would cause exceptional behavior if certain browser constants weren't available.

In this case, the test for `console` support was explicitly testing the `window` (global) namespace.  In Node, for example, the global namespace is called `global`, and there is no `window` value availble.  However, this is easily worked around, since `this` in an unbound function call is specifed to be the global namespace.