Update commands.js
Since CommonJS won't necessarily be loaded in a browser environment, it's worthwhile to cleanup code that would cause exceptional behavior if certain browser constants weren't available.

In this case, a DOM element was being created and tested to determine which implementation of `scale` should be used.  Since `document` (and, indeed, the DOM) aren't available in Node, this change defers these tests until the first time `scale` is called – at that point, it executes the same code as before, *redefining* the `scale` function with a specific implementation, then calling into that new function.