Fix build on macOS Big Sur
The build system has been using `sed -i` to do inline string replacement
in files, but the `-i` option to sed is not POSIX and its behavior
varies between GNU and BSD.

The build system attempted to detect which system you were on by
checking the behavior of a different flag (`-r`), but this detection no
longer works correctly on macOS Big Sur because it has added the `-r`
flag for compatibility with GNU (causing it to be incorrectly detected
as GNU).

Fix this by replacing `sed -i` with `perl -pi -e`. Perl is very widely
available, and its behavior here is more standard across systems.

Ref:
* https://stackoverflow.com/questions/3156178/how-do-i-use-perl-like-sed/37686753#37686753
* https://stackoverflow.com/questions/5694228/sed-in-place-flag-that-works-both-on-mac-bsd-and-linux/22247781#22247781

Fixes https://github.com/mathquill/mathquill/issues/925