Never use shorthand speech when issuing ARIA alerts
Alerts are fired only when the user is editing an expression. For these cases, it is vital to ensure the technical form of part of an expression is spoken to avoid ambiguity.

For instance, Tabbing out of a superscript of 2 results in Mathquill saying 'out of superscript, 2, baseline' instead of 'out of squared.'