Replace jQuery's .eq()
Note, I ended up changing the data model for delimiters here slightly.
They need to store references to their left and right sides, and they
were doing this with a jQ collection with two elements.

These two end elements typically have other elements between them, so
storing them in a DOMFragment also stores all the other elements between
them. This means that numerical indexing was working differently than it
used to.

Switch to storing these two delimiters as separate [L] and [R]
properties on an object.

Note that this means the new DOMFragment::eq() method doesn't actually
have any callers in the codebase. We could remove it, but I thnk it's ok
to just have in case we need it later.