Replace jquery's width function.
This was only being used in once place: the logic for splitting a text
block into to pieces to insert the cursor near the mouse position. This
logic was already busted in several ways, so I'm making a larger change
here than I otherwise would to replace it with something that should be
better.

Unfortunately, selection of text nodes is so busted both here and on
master that it's basically impossible to test the correctness of this
right now. I'm starting to think it might actually be easier to
reimplement the Text feature from scratch than to try to do small
refactors of code that has been broken for at least 9 years.

Ref:
* https://github.com/mathquill/mathquill/commit/3eb31d07b6b65319f3db85eb1bd2ffa02860095f#r65417348