Fix places that used .first() incorrectly.
first() returns the first Element of a fragment, which may not exist
even when the fragment has other kinds of Nodes (like Text or Comment
nodes). When moving the cursor around, we want to place it next to the
outermost Node in a fragment, no matter what that node is.