Fix "unrecognized keys that move cursor and clear selection" tests
This PR from 2017 https://github.com/desmosinc/mathquill/pull/81/commits/65b65de20bf87832d3e9caa2491d60669bb8a377
added some logic to only trigger `typedText` for non-arrow-key events.

This should have caused the "unrecognized keys..." unit tests to start
failing, because they depended on some logic within `typedText` to be
run when something after a key like `Cmd-Right`.

The tests didn't fail then, though, because `isArrowKey(e)` incorrectly
returned `false` if `e.originalEvent` was undefined, which was the case
for the JQuery-triggered events that the unit test used.

The tests in question come from this 2014 PR https://github.com/mathquill/mathquill/pull/294.
I have no idea if the issue that PR was fixing is still relevant today,
but I believe this change preserves the intent of that PR in a way
that's consistent with the 2017 `isArrowKey` fix, with one exception:

One of the tests checks that IF a browser sends a keypress event for a
nonprintable key (even though it’s not supposed to) and the textarea
contains exactly one character, then MQ will treat it as newly typed
text.  Based on Han’s comments in this PR
https://github.com/mathquill/mathquill/pull/294 and the related issue,
this is actually a known/acceptable bug in MQ for the case where Firefox
was behaving contrary to the standard. FF fixed this back in 2019,
so I think it's safe to remove the test.