Update tilde behavior (#247)
Adds a new configuration option: interpretAsSim. When true, tilde symbols set externally via the Mathfield.latex() method will be transformed to ]sim operators. When false (the default), tilde characters remain in the LaTeX but are internally changed to actual tilde entities that visually appear as nonbreaking spaces.

Tilde characters input from a keyboard will continue to be changed to \sim and \approx as before.