[CI] Handle Sauce Connect 4.8.1 writing info to sauce_is_ready
ba6b23ddae5e1f85ca0f3351887da32a4ecdb610 is mistaken, it's not that
Sauce Connect is writing to sauce_is_ready upon failure, it's that upon
failure we're upgrading to 4.8.1 which no longer leaves the readyfile
empty.

This commit clears the cache (by updating the cache key [1]) and checks
specifically for the error message we write to sauce_is_ready, instead
of assuming Sauce Connect will leave it empty.

[1]: https://circleci.com/docs/caching#clearing-cache