coordinate autoSubscript with charsThatBreakOut
autoSubscriptNumerals was overriding the behavior for
charsThatBreakOutOfSupSub.

As a result:

1. we _wouldn't_ break out of subscripts in this mode for expected
characters (reported by suzanne) --and--

2. It also made it impossible to type parentheses in a subscript,
or a subscripted subscript.

The latter might be preferable since those aren't things we can interpret in the calculator,
but we still need to be able to handle them since they're valid and pasteable latex.